+++ frontend/src/data/productosDemo.js
@@ -1,0 +1,146 @@
+export const productosDemo = [
+  {
+    idProducto: 901,
+    nombre: 'Bloque EcoBase',
+    descripcion: 'Bloque ligero de plastico reciclable para primeras hiladas y tramos medios de muro modular.',
+    tipo: 'Bloque',
+    material: 'Plastico reciclable',
+    proveedor: 'Plasticos renovables ByFusion',
+    precio: 42.5,
+    alto: 22.7,
+    ancho: 19.7,
+    largo: 39.4
+  },
+  {
+    idProducto: 902,
+    nombre: 'Bloque EcoPlano',
+    descripcion: 'Bloque de plastico reciclable con superficie plana para remates superiores y zonas bajo huecos.',
+    tipo: 'Bloque',
+    material: 'Plastico reciclable',
+    proveedor: 'Plasticos renovables ByFusion',
+    precio: 41,
+    alto: 20.3,
+    ancho: 19.7,
+    largo: 39.4
+  },
+  {
+    idProducto: 903,
+    nombre: 'Bloque EcoUnion',
+    descripcion: 'Bloque de plastico reciclable para encuentros, huecos de ventana y transiciones escalonadas.',
+    tipo: 'Bloque',
+    material: 'Plastico reciclable',
+    proveedor: 'Plasticos renovables ByFusion',
+    precio: 44,
+    alto: 22.7,
+    ancho: 19.7,
+    largo: 39.4
+  },
+  {
+    idProducto: 904,
+    nombre: 'Columna EcoStruct 120',
+    descripcion: 'Columna modular de plastico reciclable para muros, particiones y cerramientos ligeros.',
+    tipo: 'Pilar',
+    material: 'Plastico reciclable',
+    proveedor: 'Plasticos renovables ByFusion',
+    precio: 180,
+    alto: 120,
+    ancho: 39.4,
+    largo: 39.4
+  },
+  {
+    idProducto: 905,
+    nombre: 'Columna EcoStruct 180',
+    descripcion: 'Columna alta de plastico reciclable para puntos de refuerzo en cerramientos modulares.',
+    tipo: 'Pilar',
+    material: 'Plastico reciclable',
+    proveedor: 'Plasticos renovables ByFusion',
+    precio: 258,
+    alto: 180,
+    ancho: 39.4,
+    largo: 39.4
+  },
+  {
+    idProducto: 906,
+    nombre: 'Columna EcoCorner',
+    descripcion: 'Pilar de plastico reciclable pensado para esquinas, encuentros y cambios de direccion.',
+    tipo: 'Pilar',
+    material: 'Plastico reciclable',
+    proveedor: 'Plasticos renovables ByFusion',
+    precio: 195,
+    alto: 120,
+    ancho: 39.4,
+    largo: 39.4
+  },
+  {
+    idProducto: 907,
+    nombre: 'Bloque H80 Max',
+    descripcion: 'Bloque de hormigon de gran formato para muros de contencion de tierras y cargas exigentes.',
+    tipo: 'Bloque',
+    material: 'Hormigon',
+    proveedor: 'Hormigon Forpol Group',
+    precio: 152,
+    alto: 80,
+    ancho: 80,
+    largo: 160
+  },
+  {
+    idProducto: 908,
+    nombre: 'Pilar H80 Refuerzo',
+    descripcion: 'Pilar de hormigon pesado para refuerzo vertical en muros de contencion y zonas de carga.',
+    tipo: 'Pilar',
+    material: 'Hormigon',
+    proveedor: 'Hormigon Forpol Group',
+    precio: 210,
+    alto: 160,
+    ancho: 80,
+    largo: 80
+  },
+  {
+    idProducto: 909,
+    nombre: 'Bloque H60 Max',
+    descripcion: 'Bloque de hormigon para separadores de aridos, desechos y materiales a granel.',
+    tipo: 'Bloque',
+    material: 'Hormigon',
+    proveedor: 'Hormigon Forpol Group',
+    precio: 168,
+    alto: 60,
+    ancho: 60,
+    largo: 240
+  },
+  {
+    idProducto: 910,
+    nombre: 'Pilar H60 Modular',
+    descripcion: 'Pilar de hormigon para apoyo intermedio en separadores de materiales y muros industriales.',
+    tipo: 'Pilar',
+    material: 'Hormigon',
+    proveedor: 'Hormigon Forpol Group',
+    precio: 148,
+    alto: 120,
+    ancho: 60,
+    largo: 60
+  },
+  {
+    idProducto: 911,
+    nombre: 'Bloque H40 Cerramiento',
+    descripcion: 'Bloque de hormigon para vallas y cerramientos perimetrales sin contencion de tierras.',
+    tipo: 'Bloque',
+    material: 'Hormigon',
+    proveedor: 'Hormigon Forpol Group',
+    precio: 112,
+    alto: 80,
+    ancho: 40,
+    largo: 160
+  },
+  {
+    idProducto: 912,
+    nombre: 'Pilar H40 Cerramiento',
+    descripcion: 'Pilar de hormigon para esquinas y remates en vallas y cerramientos perimetrales.',
+    tipo: 'Pilar',
+    material: 'Hormigon',
+    proveedor: 'Hormigon Forpol Group',
+    precio: 98,
+    alto: 120,
+    ancho: 40,
+    largo: 80
+  }
+]
