+++ frontend/src/styles/galeria.css
@@ -1,0 +1,358 @@
+﻿/* ============================================================================
+   PÃGINA GALERÃA - Proyectos con filtros, cards con overlays y proyecto destacado
+   - gallery-page: Contenedor principal
+   - gallery-intro-card: Card de introducciÃ³n con imagen de fondo
+   - gallery-project-card: Cards de proyectos con overlay degradado
+   - gallery-feature-card: Proyecto destacado (grande) con contenido absoluto
+   - gallery-toolbar: Filtros y ordenamiento
+   - gallery-favorite-btn: BotÃ³n de favoritos en esquina
+   ============================================================================ */
+
+.gallery-intro-card,
+.gallery-project-card,
+.gallery-feature-card,
+.gallery-cta-card {
+  border-color: rgba(23, 63, 95, 0.18);
+  border-radius: 8px;
+  box-shadow: 0 14px 32px rgba(15, 23, 42, 0.06);
+  overflow: hidden;
+}
+
+.gallery-intro-card {
+  height: 300px;
+  min-height: 300px;
+  margin-bottom: 22px;
+}
+
+.gallery-intro-card .row,
+.gallery-feature-card .row {
+  min-height: inherit;
+}
+
+.gallery-intro-card .row {
+  height: 100%;
+}
+
+.gallery-intro-card [class^="col-"],
+.gallery-feature-card [class^="col-"] {
+  display: flex;
+}
+
+.gallery-intro-card [class^="col-"] {
+  height: 100%;
+}
+
+.gallery-intro-copy {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+  height: 100%;
+  padding: 24px clamp(28px, 5vw, 70px);
+  text-align: center;
+}
+
+.gallery-intro-copy h1,
+.gallery-feature-copy h2,
+.gallery-cta-card h2 {
+  margin: 0;
+  color: #0f3d63;
+  font-weight: 800;
+}
+
+.gallery-intro-copy p,
+.gallery-feature-copy p,
+.gallery-cta-card p {
+  margin: 10px 0 0;
+  color: #173f5f;
+}
+
+.gallery-intro-media,
+.gallery-feature-media,
+.gallery-project-media {
+  background: #8d969d;
+  color: #fff;
+  font-weight: 700;
+  overflow: hidden;
+}
+
+.gallery-image-panel {
+  background-color: #8d969d;
+  background-position: center;
+  background-size: cover;
+  background-repeat: no-repeat;
+}
+
+.gallery-image-panel-intro {
+  background-image: var(--gallery-intro-image, none);
+}
+
+.gallery-image-panel-feature {
+  background-image: var(--gallery-feature-image, none);
+}
+
+.gallery-image-panel span {
+  color: #fff;
+  text-shadow: 0 1px 2px rgba(15, 23, 42, 0.36);
+}
+
+.gallery-cover-image {
+  display: block;
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+}
+
+.gallery-intro-media {
+  position: relative;
+  width: 100%;
+  min-height: 300px;
+  height: 100%;
+}
+
+.gallery-intro-media .gallery-cover-image {
+  object-position: center center;
+}
+
+.gallery-intro-media::before {
+  content: "";
+  position: absolute;
+  inset: 0;
+  background: linear-gradient(90deg, rgba(255, 255, 255, 0.86), rgba(255, 255, 255, 0.14) 42%, transparent 72%);
+  pointer-events: none;
+  z-index: 1;
+}
+
+.gallery-toolbar {
+  margin-bottom: 22px;
+}
+
+.gallery-filter-btn,
+.gallery-outline-btn,
+.gallery-cta-btn {
+  min-width: 132px;
+  border: 1px solid #111;
+  border-radius: 5px;
+  background: #fff;
+  color: #0b0b0b;
+  font-size: 0.82rem;
+  text-transform: uppercase;
+}
+
+.gallery-filter-btn {
+  min-width: 120px;
+}
+
+.gallery-filter-btn:hover,
+.gallery-outline-btn:hover,
+.gallery-cta-btn:hover {
+  background: #f4f8ff;
+  border-color: #111;
+}
+
+.gallery-sort-select {
+  max-width: 220px;
+  border-color: #111;
+  color: #173f5f;
+}
+
+.gallery-grid {
+  margin-bottom: 24px;
+}
+
+.gallery-project-media {
+  position: relative;
+  aspect-ratio: 16 / 6.2;
+  min-height: 230px;
+}
+
+.gallery-project-card .card-body {
+  padding: 18px 18px;
+}
+
+.gallery-project-card h2,
+.gallery-project-card p {
+  margin: 0;
+  color: inherit;
+}
+
+.gallery-project-card h2 {
+  font-size: 1rem;
+  font-weight: 800;
+}
+
+.gallery-project-card p {
+  margin-top: 5px;
+  font-size: 0.85rem;
+}
+
+.gallery-project-description {
+  max-width: calc(100% - 54px);
+  line-height: 1.45;
+}
+
+.gallery-project-card .gallery-area {
+  margin-top: 11px;
+  color: #fff;
+}
+
+.gallery-card-overlay {
+  position: absolute;
+  inset-inline: 0;
+  bottom: 0;
+  min-height: 122px;
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-end;
+  color: #fff;
+  background: linear-gradient(180deg, transparent 0%, rgba(5, 18, 26, 0.58) 34%, rgba(5, 18, 26, 0.92) 100%);
+  text-shadow: 0 1px 2px rgba(5, 18, 26, 0.42);
+}
+
+.gallery-favorite-btn,
+.gallery-card-action {
+  position: absolute;
+  z-index: 2;
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  border: 1px solid rgba(255, 255, 255, 0.76);
+  background: rgba(255, 255, 255, 0.88);
+  color: #173f5f;
+  line-height: 1;
+}
+
+.gallery-favorite-btn {
+  top: 12px;
+  right: 12px;
+  width: 34px;
+  height: 34px;
+  border-radius: 999px;
+  font-size: 1rem;
+}
+
+.gallery-card-action {
+  right: 14px;
+  bottom: 18px;
+  width: 38px;
+  height: 38px;
+  border-radius: 999px;
+  font-size: 1.05rem;
+  font-weight: 800;
+}
+
+.gallery-favorite-btn:hover,
+.gallery-card-action:hover {
+  background: #fff;
+  border-color: #fff;
+  color: #0f3d63;
+}
+
+.gallery-feature-card {
+  position: relative;
+  min-height: 720px;
+  margin-bottom: 28px;
+}
+
+.gallery-feature-copy {
+  position: relative;
+  z-index: 2;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  width: min(100%, 520px);
+  min-height: inherit;
+  padding: 42px clamp(28px, 5vw, 64px);
+  color: #fff;
+  background: linear-gradient(90deg, rgba(5, 18, 26, 0.88), rgba(5, 18, 26, 0.62) 68%, transparent);
+}
+
+.gallery-feature-copy h2,
+.gallery-feature-copy p {
+  color: #fff;
+  text-shadow: 0 1px 2px rgba(5, 18, 26, 0.42);
+}
+
+.gallery-badge {
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  width: min(100%, 520px);
+  min-width: 0;
+  min-height: 34px;
+  margin-bottom: 28px;
+  border: 1px solid #111;
+  border-radius: 5px;
+  background: #fff;
+  color: #0b0b0b;
+  font-size: 0.78rem;
+  font-weight: 700;
+  text-transform: uppercase;
+}
+
+.gallery-area {
+  font-size: 0.85rem;
+  font-weight: 700;
+}
+
+.gallery-feature-architect {
+  font-weight: 700;
+}
+
+.gallery-feature-actions {
+  margin-top: 26px;
+}
+
+.gallery-feature-image {
+  position: absolute;
+  inset: 0;
+  width: 100%;
+  height: 100%;
+}
+
+.gallery-feature-card::after {
+  content: "";
+  position: absolute;
+  inset: 0;
+  background:
+    linear-gradient(180deg, rgba(5, 18, 26, 0.12), rgba(5, 18, 26, 0.42)),
+    linear-gradient(90deg, rgba(5, 18, 26, 0.64), transparent 58%);
+  pointer-events: none;
+  z-index: 1;
+}
+
+.gallery-feature-favorite {
+  top: 18px;
+  right: 18px;
+  z-index: 3;
+}
+
+.gallery-feature-action-btn {
+  border-color: rgba(255, 255, 255, 0.86);
+  background: rgba(255, 255, 255, 0.12);
+  color: #fff;
+}
+
+.gallery-feature-action-btn:hover {
+  background: #fff;
+  color: #0f3d63;
+}
+
+.gallery-pagination {
+  margin: 8px 0 24px;
+}
+
+.gallery-cta-card {
+  min-height: 150px;
+}
+
+.gallery-cta-card .card-body {
+  padding: 42px clamp(28px, 5vw, 70px);
+}
+
+.gallery-cta-btn {
+  min-width: 230px;
+  min-height: 50px;
+}
+
