+++ frontend/src/styles/legacy-pages.css
@@ -1,0 +1,439 @@
+﻿.register-shell {
+  padding-top: 20px;
+  padding-bottom: 40px;
+}
+
+.register-container {
+  min-height: calc(100svh - 160px);
+  display: flex;
+  align-items: center;
+}
+
+.register-row {
+  width: 100%;
+  max-width: 820px;
+}
+
+.register-card {
+  border-radius: 1rem;
+  width: 100%;
+}
+
+.register-card .form-outline {
+  text-align: left;
+}
+
+.register-card .form-label {
+  display: block;
+  margin-top: 0.35rem;
+  margin-bottom: 0;
+}
+
+.register-card .form-control-lg {
+  min-height: 3.1rem;
+}
+
+.login-shell {
+  padding-top: 20px;
+  padding-bottom: 40px;
+}
+
+.login-container {
+  min-height: calc(100svh - 160px);
+  display: flex;
+  align-items: center;
+}
+
+.login-row {
+  width: 100%;
+  max-width: 860px;
+}
+
+.login-card {
+  border-radius: 1rem;
+  width: 100%;
+}
+
+.login-card .form-outline {
+  text-align: left;
+}
+
+.login-card .form-label {
+  display: block;
+  margin-top: 0.35rem;
+  margin-bottom: 0;
+}
+
+.login-card .form-control-lg {
+  min-height: 3.1rem;
+}
+
+.login-footer {
+  margin-top: 0.5rem;
+}
+
+.perfil-shell {
+  padding-top: 20px;
+  padding-bottom: 40px;
+}
+
+.perfil-container {
+  min-height: calc(100svh - 160px);
+  display: flex;
+  align-items: center;
+}
+
+.perfil-card {
+  border-radius: 1rem;
+  width: 100%;
+  max-width: 100%;
+  box-shadow: 0 18px 40px rgba(0, 0, 0, 0.24);
+}
+
+.perfil-note {
+  max-width: 42rem;
+  margin-left: auto;
+  margin-right: auto;
+  line-height: 1.55;
+}
+
+.perfil-table-wrap {
+  border: 1px solid rgba(255, 255, 255, 0.12);
+  border-radius: 0.9rem;
+  overflow: hidden;
+}
+
+.perfil-table {
+  margin-bottom: 0;
+  table-layout: fixed;
+}
+
+.perfil-table th,
+.perfil-table td {
+  padding: 1rem 1rem;
+  border-color: rgba(255, 255, 255, 0.12) !important;
+  vertical-align: middle;
+}
+
+.perfil-table th {
+  width: 45%;
+  color: rgba(255, 255, 255, 0.72);
+  font-weight: 700;
+  letter-spacing: 0.02em;
+}
+
+.perfil-table td {
+  color: #fff;
+  text-align: right;
+  word-break: break-word;
+}
+
+.perfil-table .badge {
+  font-size: 0.8rem;
+  letter-spacing: 0.03em;
+}
+
+.perfil-actions {
+  flex-wrap: wrap;
+}
+
+.perfil-actions .btn {
+  min-width: 11rem;
+}
+
+.page-hero,
+.page-block,
+.page-card {
+  border-radius: 14px;
+  border: 1px solid rgba(255, 255, 255, 0.08);
+}
+
+.page-hero {
+  min-height: 250px;
+  background: linear-gradient(180deg, #2a2f36 0%, #1e2228 100%);
+  color: #f8fafc;
+  display: grid;
+  place-items: center;
+  text-align: center;
+  padding: 32px;
+  margin-bottom: 24px;
+}
+
+.page-hero h1 {
+  font-size: clamp(2.2rem, 5vw, 4.3rem);
+  margin: 0;
+  color: #f8fafc;
+}
+
+.page-hero p {
+  margin-top: 14px;
+  max-width: 56ch;
+  color: rgba(248, 250, 252, 0.82);
+}
+
+.page-grid {
+  display: grid;
+  gap: 18px;
+  grid-template-columns: repeat(auto-fit, minmax(min(100%, 240px), 1fr));
+  width: 100%;
+}
+
+.page-card {
+  background: #23272d;
+  color: #fff;
+  overflow: hidden;
+  min-height: 220px;
+  box-shadow: 0 12px 28px rgba(15, 23, 42, 0.12);
+}
+
+.compact-card {
+  min-height: auto;
+}
+
+.modular-media {
+  background:
+    linear-gradient(135deg, rgba(126, 137, 148, 0.9), rgba(70, 82, 92, 0.95));
+  color: #f8fafc;
+  font-weight: 800;
+  text-transform: uppercase;
+}
+
+.page-card-media {
+  min-height: 150px;
+  display: grid;
+  place-items: center;
+  background: #8a929d;
+  color: #20252d;
+  font-size: 1rem;
+}
+
+.page-card-body {
+  padding: 14px 16px 16px;
+  text-align: left;
+}
+
+.page-card-body h2,
+.page-card-body h3,
+.page-card-body p {
+  margin: 0;
+}
+
+.page-card-body p {
+  margin-top: 8px;
+  color: rgba(255, 255, 255, 0.84);
+}
+
+.page-actions {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 10px;
+  justify-content: center;
+  margin-top: 24px;
+}
+
+.catalog-search-row {
+  justify-content: flex-start;
+}
+
+.catalog-back-button {
+  white-space: nowrap;
+}
+
+.page-link-button {
+  padding: 0.7rem 1rem;
+  border-radius: 10px;
+  border: 1px solid rgba(255, 255, 255, 0.18);
+  color: #fff;
+}
+
+.page-link-button.is-active,
+.page-link-button:hover {
+  background: rgba(255, 255, 255, 0.12);
+}
+
+.home-card-button {
+  width: 100%;
+  text-align: left;
+  padding: 0;
+  cursor: pointer;
+}
+
+.product-link-block,
+.products-intro,
+.design-board,
+.cart-summary {
+  margin-top: 28px;
+  border-radius: 14px;
+  border: 1px solid rgba(255, 255, 255, 0.1);
+  background: #20252b;
+  color: #fff;
+  padding: 24px;
+  box-shadow: 0 12px 28px rgba(15, 23, 42, 0.12);
+}
+
+.product-link-block,
+.products-intro,
+.cart-summary {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: 18px;
+  text-align: left;
+}
+
+.product-link-block h2,
+.products-intro h2,
+.design-panel h2,
+.cart-summary h2 {
+  margin: 0;
+  color: #fff;
+}
+
+.product-link-block p,
+.products-intro p,
+.cart-summary p {
+  margin-top: 8px;
+  max-width: 65ch;
+  color: rgba(255, 255, 255, 0.78);
+}
+
+.cart-total {
+  display: grid;
+  gap: 4px;
+  min-width: 12rem;
+  border-radius: 12px;
+  background: rgba(255, 255, 255, 0.08);
+  padding: 1rem;
+  text-align: right;
+}
+
+.cart-total span {
+  color: rgba(255, 255, 255, 0.68);
+}
+
+.cart-total strong {
+  color: #fff;
+  font-size: 1.5rem;
+}
+
+.product-specs {
+  display: grid;
+  gap: 8px;
+  min-width: 15rem;
+}
+
+.product-specs span {
+  border-radius: 10px;
+  background: rgba(255, 255, 255, 0.08);
+  padding: 0.55rem 0.75rem;
+  color: rgba(255, 255, 255, 0.86);
+}
+
+.catalog-grid {
+  scroll-margin-top: 120px;
+}
+
+.auth-container,
+.profile-container,
+.usuarios-container {
+  width: 100%;
+}
+
+.auth-card,
+.perfil-card,
+.usuarios-card {
+  border: 1px solid rgba(255, 255, 255, 0.1);
+}
+
+.usuarios-card {
+  overflow: hidden;
+}
+
+.design-board {
+  display: grid;
+  grid-template-columns: minmax(220px, 300px) 1fr;
+  gap: 20px;
+}
+
+.design-panel {
+  text-align: left;
+}
+
+.tool-stack {
+  display: grid;
+  gap: 10px;
+  margin-top: 18px;
+}
+
+.tool-stack button {
+  border: 1px solid rgba(255, 255, 255, 0.18);
+  border-radius: 10px;
+  background: rgba(255, 255, 255, 0.06);
+  color: #fff;
+  padding: 0.7rem 0.9rem;
+  text-align: left;
+}
+
+.plan-preview {
+  display: grid;
+  grid-template-columns: repeat(6, minmax(30px, 1fr));
+  gap: 8px;
+  padding: 16px;
+  border-radius: 12px;
+  background:
+    linear-gradient(rgba(255, 255, 255, 0.05) 1px, transparent 1px),
+    linear-gradient(90deg, rgba(255, 255, 255, 0.05) 1px, transparent 1px),
+    #161a1f;
+  background-size: 32px 32px;
+}
+
+.plan-preview span {
+  aspect-ratio: 1;
+  border-radius: 8px;
+  border: 1px solid rgba(255, 255, 255, 0.12);
+  background: rgba(255, 255, 255, 0.04);
+}
+
+.plan-preview span.is-filled {
+  background: #45a66b;
+  box-shadow: 0 0 0 2px rgba(69, 166, 107, 0.18);
+}
+
+.step-number {
+  display: inline-grid;
+  place-items: center;
+  width: 2rem;
+  height: 2rem;
+  margin-bottom: 0.8rem;
+  border-radius: 999px;
+  background: #45a66b;
+  color: #fff;
+  font-weight: 800;
+}
+
+.home-card-button:focus-visible,
+.nav-button:focus-visible,
+.page-link-button:focus-visible,
+.brand-button:focus-visible {
+  outline: 2px solid #93c5fd;
+  outline-offset: 2px;
+}
+
+.navbar-nav {
+    --bs-nav-link-padding-x: 0;
+    --bs-nav-link-padding-y: 0.5rem;
+    --bs-nav-link-font-weight: ;
+    --bs-nav-link-color: var(--bs-navbar-color);
+    --bs-nav-link-hover-color: var(--bs-navbar-hover-color);
+    --bs-nav-link-disabled-color: var(--bs-navbar-disabled-color);
+    display: flex;
+    flex-direction: row;
+    padding-left: 0;
+    margin-bottom: 0;
+    list-style: none;
+}
+
+.gradient-custom {
+  background: #ffffff;
+}
+
