+++ frontend/src/styles/site-footer.css
@@ -1,0 +1,57 @@
+﻿.site-footer {
+  width: 100%;
+}
+
+.site-benefits-footer {
+  display: grid;
+  grid-template-columns: repeat(4, minmax(0, 1fr));
+  gap: 18px;
+  width: calc(100% - 40px);
+  margin: 34px auto 10px;
+  border-radius: 12px;
+  background: #d9ecff;
+  padding: 18px 24px;
+}
+
+.site-benefits-footer article {
+  display: flex;
+  gap: 12px;
+  align-items: center;
+  justify-content: center;
+}
+
+.site-benefits-footer span {
+  color: #0f3d63;
+  font-size: 1.6rem;
+  font-weight: 800;
+}
+
+.site-benefits-footer h2,
+.site-benefits-footer p {
+  margin: 0;
+}
+
+.site-benefits-footer h2 {
+  color: #0f172a;
+  font-size: 0.98rem;
+  font-weight: 850;
+}
+
+.site-benefits-footer p {
+  margin-top: 2px;
+  color: #475569;
+  font-size: 0.86rem;
+}
+
+.site-footer-copy {
+  width: calc(100% - 40px);
+  margin: 18px auto;
+  border-radius: 10px;
+  background: #d9efd8;
+  color: #173f5f;
+  font-size: 0.82rem;
+  font-weight: 700;
+  padding: 12px 18px;
+  text-align: center;
+}
+
