+++ frontend/src/utils/validators.js
@@ -1,0 +1,1 @@
+export const isValidEmail = (email) => /^[^\s@]+@[^\s@]+\.[^\s@]+$/.test(email)
